Move video classes from MainGazeTracker into Video
