+++ MainGazeTracker.h
@@ -1,11 +1,10 @@
 #pragma once
 
-#include <opencv/highgui.h>
-
 #include "TrackingSystem.h"
 #include "Calibrator.h"
 #include "GameWindow.h"
 #include "OutputMethods.h"
+#include "Video.h"
 
 struct CommandLineArguments {
 	std::vector<std::string> parameters;
@@ -23,36 +22,6 @@ struct Command {
 	std::string commandName;
 
 	Command(long number, std::string name);
-};
-
-class VideoInput {
-private:
-	CvCapture *_capture;
-	long _lastFrameTime;
-
-public:
-	int frameCount;
-	IplImage *frame;
-	CvSize size;
-	bool captureFromVideo;
-	std::string resolutionParameter;
-
-	VideoInput();
-	VideoInput(std::string resolution);
-	VideoInput(std::string resolution, std::string filename, bool dummy);
-	~VideoInput();
-	void updateFrame();
-	double getResolution();
-};
-
-class VideoWriter {
-public:
-	VideoWriter(CvSize size, std::string filename);
-	~VideoWriter();
-	void write(const IplImage *image);
-
-private:
-	CvVideoWriter *_video;
 };
 
 /* class FileInput; */
